asm-verify #194 Wave B: refactor 8 free fns to the binary signature (pair)
8 ported free functions paired under a different signature than the binary, so
asm-verify still couldn't diff them. Refactored to the binary's exact symbol
(all binary-faithful, behavior-preserving):
- StringHash: two-overload -- out-of-line StringHash(const char*, int len) (the
  int is the byte length, per binary @0x00252a10) + inline 1-arg forwarder.
  Emits both _Z10StringHashPKci and _Z10StringHashPKc; 229 call sites unchanged.
- GameInitialise: restore (void* window, const char* config) @0x0011d22c (args
  unused -- SDL owns its window; // DIFFERS); callers pass nullptr,nullptr.
- CreateEntity: int -> long @0x001d8fec (+ TODO for the unported case-6 FruitRay).
- GetBonusTotal: drop static, uint64_t -> unsigned long @0x0012e4c4.
- SetCurrentModeHighscore: free bool(int) wrapper @0x00119f24 over the save
  singleton (member kept).
- MoveFruitZPositionToBack: de-method, Fruit* -> float& @0x001ca674.
- GetComboStarText/Texture: de-method + uint8_t -> COMBO_TYPE enum (25 values)
  @0x001325f8 / @0x00132a94.
- IsFastHardware: drop static @0x0011f394.
- CriticalFlash: already free + const-ref -> pairs as-is (verify only).
AddSlice deferred to the faithful SliceEffect subsystem port (#196), not a rename.
Build green; 20/20 tests pass.

+++ src/entities/EntityFactory.cpp
@@ -11,7 +11,7 @@
 
 #include <cstdio>
 
-// CreateEntity — binary-faithful port of 0x001d90f4 (v1.6.1).
+// CreateEntity — binary-faithful port of _Z12CreateEntityl @0x001d8fec (v1.6.1).
 //
 // Binary pseudocode:
 //   Mortar::Entity* CreateEntity(long entityType) {
@@ -22,13 +22,12 @@
 //       case 3:  return new SlashEntity();  // 0x188 bytes
 //       case 4:  return new BombBlast();    // 0x70  bytes
 //       case 5:  return new Jiblet();       // 0xB0  bytes (memset 0xB0 first)
+//       case 6:  return new FruitRay();     // (unported)
 //       default: return nullptr;
 //       }
 //   }
-//
-// Analysed: 2026-04-23T01:30
 
-Mortar::Entity* CreateEntity(int entityType) {
+Mortar::Entity* CreateEntity(long entityType) {
     switch (entityType) {
     case 0:  return new Fruit();
     case 1:  return new Bomb();
@@ -36,8 +35,9 @@ Mortar::Entity* CreateEntity(int entityType) {
     case 3:  return new SlashEntity();
     case 4:  return new BombBlast();
     case 5:  return new Jiblet();
+    // TODO: v1.6.1 CreateEntity @0x001d8fec case 6 FruitRay (unported)
     default:
-        LOG_WARN("ENTITYFACTORY", "unknown entity type %d", entityType);
+        LOG_WARN("ENTITYFACTORY", "unknown entity type %ld", entityType);
         return nullptr;
     }
 }
